Rename keyboard implementation filenames
