+++ keyboards/aos/tkl/tkl.c
@@ -15,4 +15,4 @@ You should have received a copy of the GNU General Public License
 along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-#include "ace_of_spades.h"
+#include "tkl.h"
